rename lbrynet.lbryfile to lbrynet.lbry_file
