+++ lbrynet/lbry_file/EncryptedFileMetadataManager.py
