+++ lbrynet/lbry_file/StreamDescriptor.py
