+++ lbrynet/lbry_file/__init__.py
@@ -1,0 +1,2 @@
+from lbrynet.lbry_file.StreamDescriptor import get_sd_info
+from lbrynet.lbry_file.StreamDescriptor import publish_sd_blob
