+++ lbrynet/lbry_file/client/EncryptedFileMetadataHandler.py
