+++ lbrynet/lbry_file/client/EncryptedFileOptions.py
@@ -1,5 +1,5 @@
-from lbrynet.lbryfile.StreamDescriptor import EncryptedFileStreamType
-from lbrynet.lbryfile.StreamDescriptor import EncryptedFileStreamDescriptorValidator
+from lbrynet.lbry_file.StreamDescriptor import EncryptedFileStreamType
+from lbrynet.lbry_file.StreamDescriptor import EncryptedFileStreamDescriptorValidator
 from lbrynet.core.DownloadOption import DownloadOption, DownloadOptionChoice
