+++ lbrynet/lbryfilemanager/EncryptedFileDownloader.py
@@ -10,11 +10,11 @@ from lbrynet.core.client.StreamProgressManager import FullStreamProgressManager
 from lbrynet.core.Error import NoSuchSDHash, NoSuchStreamHash
 from lbrynet.core.utils import short_hash
 from lbrynet.core.StreamDescriptor import StreamMetadata
-from lbrynet.lbryfile.client.EncryptedFileDownloader import EncryptedFileSaver
-from lbrynet.lbryfile.client.EncryptedFileDownloader import EncryptedFileDownloader
+from lbrynet.lbry_file.client.EncryptedFileDownloader import EncryptedFileSaver
+from lbrynet.lbry_file.client.EncryptedFileDownloader import EncryptedFileDownloader
 from lbrynet.lbryfilemanager.EncryptedFileStatusReport import EncryptedFileStatusReport
 from lbrynet.interfaces import IStreamDownloaderFactory
-from lbrynet.lbryfile.StreamDescriptor import save_sd_info
+from lbrynet.lbry_file.StreamDescriptor import save_sd_info
 from lbrynet.core.Wallet import ClaimOutpoint
 
 log = logging.getLogger(__name__)
