+++ lbrynet/lbrynet_daemon/Publisher.py
@@ -6,7 +6,7 @@ from twisted.internet import defer
 
 from lbrynet.core import file_utils
 from lbrynet.lbryfilemanager.EncryptedFileCreator import create_lbry_file
-from lbrynet.lbryfile.StreamDescriptor import publish_sd_blob
+from lbrynet.lbry_file.StreamDescriptor import publish_sd_blob
 
 
 log = logging.getLogger(__name__)
