+++ tests/unit/lbrynet_daemon/test_Downloader.py
@@ -9,7 +9,7 @@ from lbryschema.claim import ClaimDict
 from lbrynet.core import Session, PaymentRateManager, Wallet
 from lbrynet.lbrynet_daemon import Downloader
 from lbrynet.core.StreamDescriptor import StreamDescriptorIdentifier,StreamMetadata
-from lbrynet.lbryfile.client.EncryptedFileOptions import add_lbry_file_to_sd_identifier
+from lbrynet.lbry_file.client.EncryptedFileOptions import add_lbry_file_to_sd_identifier
 from lbrynet.core.HashBlob import TempBlob
 from lbrynet.core.BlobManager import TempBlobManager
 from lbrynet.lbryfilemanager.EncryptedFileDownloader import ManagedEncryptedFileDownloaderFactory
